+++ src/app/project/initProject.js
@@ -1,6 +1,6 @@
 import Path from "path";
 import { ActionCreators } from "redux-undo";
-import { ipcRenderer, clipboard, webFrame} from "electron";
+import { ipcRenderer, webFrame } from "electron";
 import settings from "electron-settings";
 import debounce from "lodash/debounce";
 import mapValues from "lodash/mapValues";